pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "golden-rule-compiler"
version = "0.1.0"
description = "BFCL-first Golden Rule Compiler MVP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.2",
  "jsonschema>=4.23.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "orjson>=3.10.7",
]

[project.scripts]
grc = "grc.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

